feat(files): organized components in files
