+++ src/components/Footer/index.tsx
@@ -1,10 +1,8 @@
-import { CodarmeCommunity } from "./CodarCom"
-
 export const Footer = () =>{
     return(
         <footer className="h-20 w-full  flex items-center justify-end pr-10  border-t border-bastille-800 lg:mx-auto container" >
                 
-            <CodarmeCommunity width={140}/>
+         
           
         </footer>
     )
